+++ Weight Tracker/ViewModels/MonthlyStatisticsViewModel.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Weight_Tracker.DbConnection;
+using Weight_Tracker.DatabaseServices;
 using Weight_Tracker.Models;
 
 namespace Weight_Tracker.ViewModels
@@ -12,7 +12,7 @@ namespace Weight_Tracker.ViewModels
         private readonly DateTime _yearAgo = Date.YearAgo;
         private readonly DateTime _today = Date.Today;        
         
-        public ChartValues<double>? MonthlyWeightValues { get; set; }
+        public ChartValues<float>? MonthlyWeightValues { get; set; }
         public string[]? MonthAndYear { get; set; }
         public Func<double, string>? BarFormatter { get; private set; }
 
@@ -20,7 +20,7 @@ namespace Weight_Tracker.ViewModels
         {
             IEnumerable<DailyWeight> weight = new LoadDataFromDB(_yearAgo, _today).WeightStatistics;
 
-            MonthlyWeightValues = new ChartValues<double>(weight.Select(w => w.Weight));
+            MonthlyWeightValues = new ChartValues<float>(weight.Select(w => w.Weight));
             MonthAndYear = weight.Select(c => (c.Date).ToString("dd.MM.yy")).ToArray();
             BarFormatter = value => value.ToString("F1");
         }
